Serial test for tunnel cmd: moved from TestFunctional
